+++ database/migrations/2022_06_11_115838_create_document_subject_table.php
@@ -14,8 +14,8 @@ class CreateDocumentSubjectTable extends Migration
     public function up()
     {
         Schema::create('document_subject', function (Blueprint $table) {
-            $table->unsignedInteger('document_id');
-            $table->unsignedInteger('subject_id');
+            $table->unsignedBigInteger('document_id');
+            $table->unsignedBigInteger('subject_id');
             $table->timestamps();
 
             $table->primary(['document_id', 'subject_id']);
